Override git config dynamically from the command line
Starting git 1.8, it is possible to pass the following arguments:

-c name=value

to override any setting in the git config during the invocation of the
command. This commit allows passing any arbitrary of overides to the git
invocation within the module. This allows for very powerful
customizations of git without requiring the module to explicitly support
them.

Also, clean a little bit the code that relates to building the git
command. Created a git command builder class that helps with the
building of the arguments of a git command and will provide some help
injecting the overrides.

The builder, besides helping with the change also cleans up the code and
reduces the amount of repetition (i.e. git_path variable).